Report shelled out command errors
Various linters don't adhere to one strict methodology on when to return
`0` or `1` or `2` as exit codes.
For example, RuboCop returns `1` when there are violations and `2` when
any other problem is encountered. However, JSHint uses `2` when
violations are found/reported.

So, instead if there is output, but no violations could be parsed, we
can assume that something else went wrong and thus report that output as
an error.